pyproject.toml
[tool.poetry]
name = "sm-user"
version = "0.1.0"
description = ""
authors = ["Alessandro Cere"]

[tool.poetry.dependencies]
python = "^3.9"
"aws-cdk.aws-sagemaker" = "^1.93.0"
"aws-cdk.aws-ec2" = "^1.93.0"
python-dotenv = "^0.15.0"
"aws-cdk.aws-efs" = "^1.93.0"
"aws-cdk.aws-servicecatalog" = "^1.94.1"
"aws-cdk.aws-s3-deployment" = "^1.94.1"
"aws-cdk.aws-lambda-python" = "^1.95.1"

[tool.poetry.dev-dependencies]
black = {version = "^20.8b1", allow-prereleases = true}
boto3 = "^1.17.28"
isort = "^5.7.0"
GitPython = "^3.1.14"
ipykernel = "^5.5.3"
cfn-response = "^0.0.4"

[build-system]
requires = ["setuptools", "poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
